requirements.txt
fastapi
uvicorn
requests
pandas
tqdm
python-dotenv
